refactor: pure ESM package
