+++ package.json
@@ -3,6 +3,7 @@
   "version": "4.1.0",
   "description": "IoC (Indicator of Compromise) extractor",
   "main": "dist/src/index.js",
+  "type": "module",
   "types": "dist/src/index.d.ts",
   "scripts": {
     "prepare": "npm run build",
@@ -11,7 +12,7 @@
     "test": "jest",
     "coverage": "jest --coverage",
     "lint": "eslint src --ext .ts",
-    "docs:build": "typedoc",
+    "docs": "typedoc",
     "postinstall": "husky install",
     "prepublishOnly": "pinst --disable",
     "postpublish": "pinst --enable"
@@ -58,38 +59,37 @@
   "homepage": "https://github.com/ninoseki/ioc-extractor",
   "devDependencies": {
     "@types/array-unique": "0.3.0",
-    "@types/benchmark": "2.1.1",
-    "@types/jest": "28.1.4",
+    "@types/benchmark": "2.1.2",
+    "@types/jest": "29.2.0",
     "@types/memoizee": "0.4.8",
-    "@types/node": "18.0.0",
+    "@types/node": "18.11.3",
     "@types/punycode": "2.1.0",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.30.3",
-    "@typescript-eslint/parser": "5.30.3",
+    "@typescript-eslint/eslint-plugin": "5.40.1",
+    "@typescript-eslint/parser": "5.40.1",
     "benny": "3.7.1",
     "coveralls": "3.1.1",
-    "eslint": "8.19.0",
+    "eslint": "8.26.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-jest": "26.5.3",
+    "eslint-plugin-jest": "27.1.3",
     "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-simple-import-sort": "7.0.0",
+    "eslint-plugin-simple-import-sort": "8.0.0",
     "husky": "8.0.1",
-    "jest": "28.1.2",
+    "jest": "29.2.1",
     "lint-staged": "13.0.3",
     "pinst": "3.0.0",
     "prettier": "2.7.1",
-    "ts-jest": "28.0.5",
-    "ts-loader": "9.3.1",
-    "ts-node": "10.8.2",
-    "typedoc": "0.23.5",
-    "typescript": "4.7.4"
+    "ts-jest": "29.0.3",
+    "ts-loader": "9.4.1",
+    "ts-node": "10.9.1",
+    "typedoc": "0.23.17",
+    "typescript": "4.8.4"
   },
   "dependencies": {
     "array-unique": "^0.3.2",
-    "commander": "^8.3.0",
-    "get-stdin": "8.0.0",
+    "commander": "^9.4.1",
+    "get-stdin": "9.0.0",
     "memoizee": "^0.4.15",
-    "threads": "^1.7.0",
-    "uuid": "^8.3.2"
+    "uuid": "^9.0.0"
   }
 }
