+++ src/bin/ioc-extractor.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 import { program } from "commander";
-import * as getStdin from "get-stdin";
+import getStdin from "get-stdin";
 
-import { extractIOC, extractIOCAsync } from "../";
-import { convertToSTIX2 } from "../stix2/stix2";
+import { extractIOC } from "../index.js";
+import { convertToSTIX2 } from "../stix2/stix2.js";
 
 interface Options {
   stix2?: boolean;
@@ -19,7 +19,6 @@ interface Options {
 
   program
     .option("-s2, --stix2", "output in STIX2 format", false)
-    .option("-t, --threads", "use threads", false)
     .option("--disable-idn", "disable IDN extraction", false)
     .option("--disable-strict-tld", "disable strict TLD validation", false)
     .option("--disable-refang", "disable refang", false);
@@ -27,7 +26,6 @@ interface Options {
 
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
   const options = <Options>program.opts();
-  const threads = options.threads !== undefined ? options.threads : false;
   const stix2 = options.stix2 !== undefined ? options.stix2 : false;
 
   const disableIDN =
@@ -42,9 +40,7 @@ interface Options {
     options.disableRefang !== undefined ? options.disableRefang : false;
   const enableRefang = !disableRefang;
 
-  const ioc = threads
-    ? await extractIOCAsync(str, { enableIDN, strictTLD, enableRefang })
-    : extractIOC(str, { enableIDN, strictTLD, enableRefang });
+  const ioc = extractIOC(str, { enableIDN, strictTLD, enableRefang });
 
   if (stix2) {
     const stix2Obj = convertToSTIX2(ioc);
